Modify fr forces for rigid bodies with overlapping particles
Gravity force correctly computed as F= M_total *g, where M_total is the correct total mass of the body (withouth double-counting when particles overlap)
For the LJ forces, this correction is not easy to calculate. Therefore, particles are treated as additive LJ centers, the more overlapping particles you
have the greater the LJ force will be.