Improve multipole / quincke integration
Multipole and Quincke modes are independent of eachother, but can be used together.

If dipole are enabled, there are two modes : fixed or quincke.
If "fixed", we specify the body axis (x,y,z) and magnitude.
If "quincke", we only need to specify the magnitude (this requires also that Quincke mode is enabled).